refactor + add Qt PY neopixel support
I'm trying to make it so this code will run on 3 different boards
without modification. To do that, it will check board.board_id to
understand the available hardware options.

For example, on the Metro ESP32-S3, I use the red LED pin because
I've cut the neopixel power jumper to lower the deep sleep current.
For QT Py boards, there is no red LED pin, but I can turn off
neopixel power with a pin.