add battery discharge mode + refactor
This adds a battery discharge feature. To use it, you call
util.discharge(True) from the serial console, then unplug the USB
cable while the A0 jumper is connected to GND. The actual discharge
mode firmware is implemented in code.py.

The point of this is to extend the usable service life of battery
packs. Before disconnecting a battery to store a logger that is not
in active use, you can use this feature to discharge the pack to
about 3.8V.