refactor, improve admin mode + discharge
This gets rid of util.discharge() and revises the A0-GND jumpered
admin mode. Now, whenever the jumper is installed, provided the
battery voltage is not too low, the LED will send battery voltage
in Morse. This will make it easy to discharge the battery to a
reasonable storage voltage (set jumper and leave it alone for a
while) or to check the battery voltage (set jumper temporarily).